// src/client/api-back/models/CreateDocumentDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Invicta API DEV
 * Utilize essa documentação para realizar integração com o nosso sistema
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { CreateUserDocumentDto } from './CreateUserDocumentDto';
import {
    CreateUserDocumentDtoFromJSON,
    CreateUserDocumentDtoFromJSONTyped,
    CreateUserDocumentDtoToJSON,
} from './CreateUserDocumentDto';

/**
 * 
 * @export
 * @interface CreateDocumentDto
 */
export interface CreateDocumentDto {
    /**
     * Número do documento
     * @type {string}
     * @memberof CreateDocumentDto
     */
    number: string;
    /**
     * Tipo do documento
     * @type {string}
     * @memberof CreateDocumentDto
     */
    type: string;
    /**
     * Path do documento no computador
     * @type {string}
     * @memberof CreateDocumentDto
     */
    filePath: string;
    /**
     * 
     * @type {string}
     * @memberof CreateDocumentDto
     */
    status: CreateDocumentDtoStatusEnum;
    /**
     * 
     * @type {CreateUserDocumentDto}
     * @memberof CreateDocumentDto
     */
    userDocument: CreateUserDocumentDto;
}


/**
 * @export
 */
export const CreateDocumentDtoStatusEnum = {
    Sent: 'SENT',
    AwaitingVerification: 'AWAITING_VERIFICATION',
    Verified: 'VERIFIED'
} as const;
export type CreateDocumentDtoStatusEnum = typeof CreateDocumentDtoStatusEnum[keyof typeof CreateDocumentDtoStatusEnum];


/**
 * Check if a given object implements the CreateDocumentDto interface.
 */
export function instanceOfCreateDocumentDto(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "number" in value;
    isInstance = isInstance && "type" in value;
    isInstance = isInstance && "filePath" in value;
    isInstance = isInstance && "status" in value;
    isInstance = isInstance && "userDocument" in value;

    return isInstance;
}

export function CreateDocumentDtoFromJSON(json: any): CreateDocumentDto {
    return CreateDocumentDtoFromJSONTyped(json, false);
}

export function CreateDocumentDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateDocumentDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'number': json['number'],
        'type': json['type'],
        'filePath': json['filePath'],
        'status': json['status'],
        'userDocument': CreateUserDocumentDtoFromJSON(json['userDocument']),
    };
}

export function CreateDocumentDtoToJSON(value?: CreateDocumentDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'number': value.number,
        'type': value.type,
        'filePath': value.filePath,
        'status': value.status,
        'userDocument': CreateUserDocumentDtoToJSON(value.userDocument),
    };
}